gosky: move and add usage to sub-commands
