+++ cmd/gosky/debug.go
@@ -33,8 +33,8 @@ import (
 )
 
 var debugCmd = &cli.Command{
-	Name:        "debug",
-	Description: "a set of debugging utilities for atproto",
+	Name:  "debug",
+	Usage: "a set of debugging utilities for atproto",
 	Subcommands: []*cli.Command{
 		inspectEventCmd,
 		debugStreamCmd,
