+++ cmd/gosky/main.go
@@ -18,9 +18,6 @@ import (
 	"github.com/bluesky-social/indigo/api/atproto"
 	comatproto "github.com/bluesky-social/indigo/api/atproto"
 	"github.com/bluesky-social/indigo/api/bsky"
-	appbsky "github.com/bluesky-social/indigo/api/bsky"
-	"github.com/bluesky-social/indigo/atproto/identity"
-	"github.com/bluesky-social/indigo/atproto/syntax"
 	"github.com/bluesky-social/indigo/events"
 	"github.com/bluesky-social/indigo/events/schedulers/sequential"
 	lexutil "github.com/bluesky-social/indigo/lex/util"
@@ -32,7 +29,6 @@ import (
 
 	"github.com/gorilla/websocket"
 	lru "github.com/hashicorp/golang-lru"
-	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-datastore"
 	blockstore "github.com/ipfs/go-ipfs-blockstore"
 	"github.com/ipld/go-car"
@@ -82,30 +78,29 @@ func run(args []string) {
 	}
 	app.Commands = []*cli.Command{
 		actorGetSuggestionsCmd,
+		adminCmd,
 		bgsAdminCmd,
-		createSessionCmd,
 		carCmd,
+		createFeedGeneratorCmd,
+		createInviteCmd,
+		createSessionCmd,
 		debugCmd,
 		deletePostCmd,
 		didCmd,
 		feedGetCmd,
 		feedSetVoteCmd,
+		followsCmd,
+		getNotificationsCmd,
+		getRecordCmd,
+		handleCmd,
+		listAllPostsCmd,
 		newAccountCmd,
 		postCmd,
-		refreshAuthTokenCmd,
-		syncCmd,
-		listAllPostsCmd,
-		deletePostCmd,
-		getNotificationsCmd,
-		followsCmd,
-		resetPasswordCmd,
 		readRepoStreamCmd,
-		handleCmd,
-		getRecordCmd,
-		createInviteCmd,
-		adminCmd,
-		createFeedGeneratorCmd,
 		rebaseRepoCmd,
+		refreshAuthTokenCmd,
+		resetPasswordCmd,
+		syncCmd,
 	}
 
 	app.RunAndExitOnError()
@@ -182,229 +177,6 @@ var createSessionCmd = &cli.Command{
 	},
 }
 
-var postCmd = &cli.Command{
-	Name:      "post",
-	ArgsUsage: `<text>`,
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-		if err != nil {
-			return err
-		}
-
-		auth := xrpcc.Auth
-
-		text := strings.Join(cctx.Args().Slice(), " ")
-
-		resp, err := comatproto.RepoCreateRecord(context.TODO(), xrpcc, &comatproto.RepoCreateRecord_Input{
-			Collection: "app.bsky.feed.post",
-			Repo:       auth.Did,
-			Record: &lexutil.LexiconTypeDecoder{&appbsky.FeedPost{
-				Text:      text,
-				CreatedAt: time.Now().Format(util.ISO8601),
-			}},
-		})
-		if err != nil {
-			return fmt.Errorf("failed to create post: %w", err)
-		}
-
-		fmt.Println(resp.Cid)
-		fmt.Println(resp.Uri)
-
-		return nil
-	},
-}
-
-var didCmd = &cli.Command{
-	Name:  "did",
-	Flags: []cli.Flag{},
-	Subcommands: []*cli.Command{
-		didGetCmd,
-		didCreateCmd,
-		didKeyCmd,
-	},
-}
-
-var didGetCmd = &cli.Command{
-	Name:      "get",
-	ArgsUsage: `<did>`,
-	Flags: []cli.Flag{
-		&cli.BoolFlag{
-			Name:  "handle",
-			Usage: "resolve did to handle and print",
-		},
-	},
-	Action: func(cctx *cli.Context) error {
-		s := cliutil.GetDidResolver(cctx)
-
-		did := cctx.Args().First()
-
-		if cctx.Bool("handle") {
-			xrpcc, err := cliutil.GetXrpcClient(cctx, false)
-			if err != nil {
-				return err
-			}
-
-			phr := &api.ProdHandleResolver{}
-			h, _, err := api.ResolveDidToHandle(context.TODO(), xrpcc, s, phr, did)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(h)
-			return nil
-		}
-
-		doc, err := s.GetDocument(context.TODO(), did)
-		if err != nil {
-			return err
-		}
-
-		b, err := json.MarshalIndent(doc, "", "  ")
-		if err != nil {
-			return err
-		}
-
-		fmt.Println(string(b))
-		return nil
-	},
-}
-
-var didCreateCmd = &cli.Command{
-	Name:      "create",
-	ArgsUsage: `<handle> <service>`,
-	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name: "recoverydid",
-		},
-		&cli.StringFlag{
-			Name: "signingkey",
-		},
-	},
-	Action: func(cctx *cli.Context) error {
-		s := cliutil.GetPLCClient(cctx)
-
-		args, err := needArgs(cctx, "handle", "service")
-		if err != nil {
-			return err
-		}
-		handle, service := args[0], args[1]
-
-		recoverydid := cctx.String("recoverydid")
-
-		sigkey, err := cliutil.LoadKeyFromFile(cctx.String("signingkey"))
-		if err != nil {
-			return err
-		}
-
-		fmt.Println("KEYDID: ", sigkey.Public().DID())
-
-		ndid, err := s.CreateDID(context.TODO(), sigkey, recoverydid, handle, service)
-		if err != nil {
-			return err
-		}
-
-		fmt.Println(ndid)
-		return nil
-	},
-}
-
-var didKeyCmd = &cli.Command{
-	Name: "didKey",
-	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name: "keypath",
-		},
-	},
-	Action: func(cctx *cli.Context) error {
-		sigkey, err := cliutil.LoadKeyFromFile(cctx.String("keypath"))
-		if err != nil {
-			return err
-		}
-		fmt.Println(sigkey.Public().DID())
-		return nil
-	},
-}
-
-var syncCmd = &cli.Command{
-	Name: "sync",
-	Subcommands: []*cli.Command{
-		syncGetRepoCmd,
-		syncGetRootCmd,
-		syncListReposCmd,
-	},
-}
-
-var syncGetRepoCmd = &cli.Command{
-	Name:      "getRepo",
-	Usage:     "download repo from account's PDS to local file (or '-' for stdout). for hex combine with 'xxd -ps -u -c 0'",
-	ArgsUsage: `<at-identifier> [<car-file-path>]`,
-	Action: func(cctx *cli.Context) error {
-		ctx := context.Background()
-		arg := cctx.Args().First()
-		if arg == "" {
-			return fmt.Errorf("at-identifier arg is required")
-		}
-		atid, err := syntax.ParseAtIdentifier(arg)
-		if err != nil {
-			return err
-		}
-		dir := identity.DefaultDirectory()
-		ident, err := dir.Lookup(ctx, *atid)
-		if err != nil {
-			return err
-		}
-
-		carPath := cctx.Args().Get(1)
-		if carPath == "" {
-			carPath = ident.DID.String() + ".car"
-		}
-
-		xrpcc, err := cliutil.GetXrpcClient(cctx, false)
-		if err != nil {
-			return err
-		}
-		xrpcc.Host = ident.PDSEndpoint()
-		if xrpcc.Host == "" {
-			return fmt.Errorf("no PDS endpoint for identity")
-		}
-
-		log.Infof("downloading from %s to: %s", xrpcc.Host, carPath)
-		repoBytes, err := comatproto.SyncGetRepo(ctx, xrpcc, ident.DID.String(), "")
-		if err != nil {
-			return err
-		}
-
-		if carPath == "-" {
-			_, err = os.Stdout.Write(repoBytes)
-			return err
-		} else {
-			return os.WriteFile(carPath, repoBytes, 0666)
-		}
-	},
-}
-
-var syncGetRootCmd = &cli.Command{
-	Name:      "getRoot",
-	ArgsUsage: `<did>`,
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, false)
-		if err != nil {
-			return err
-		}
-
-		ctx := context.TODO()
-
-		root, err := comatproto.SyncGetHead(ctx, xrpcc, cctx.Args().First())
-		if err != nil {
-			return err
-		}
-
-		fmt.Println(root.Root)
-
-		return nil
-	},
-}
-
 func jsonPrint(i any) {
 	b, err := json.MarshalIndent(i, "", "  ")
 	if err != nil {
@@ -412,169 +184,6 @@ func jsonPrint(i any) {
 	}
 
 	fmt.Println(string(b))
-}
-
-func prettyPrintPost(p *appbsky.FeedDefs_FeedViewPost, uris bool) {
-	fmt.Println(strings.Repeat("-", 60))
-	rec := p.Post.Record.Val.(*appbsky.FeedPost)
-	fmt.Printf("%s (%s)", p.Post.Author.Handle, rec.CreatedAt)
-	if uris {
-		fmt.Println(" -- ", p.Post.Uri)
-	} else {
-		fmt.Println(":")
-	}
-	fmt.Println(rec.Text)
-}
-
-var feedGetCmd = &cli.Command{
-	Name: "feed",
-	Flags: []cli.Flag{
-		&cli.IntFlag{
-			Name:  "count",
-			Value: 100,
-		},
-		&cli.StringFlag{
-			Name:  "author",
-			Usage: "specify handle of user to list their authored feed",
-		},
-		&cli.BoolFlag{
-			Name:  "raw",
-			Usage: "print out feed in raw json",
-		},
-		&cli.BoolFlag{
-			Name:  "uris",
-			Usage: "include URIs in pretty print output",
-		},
-	},
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-		if err != nil {
-			return err
-		}
-
-		ctx := context.TODO()
-
-		raw := cctx.Bool("raw")
-
-		uris := cctx.Bool("uris")
-
-		author := cctx.String("author")
-		if author != "" {
-			if author == "self" {
-				author = xrpcc.Auth.Did
-			}
-
-			tl, err := appbsky.FeedGetAuthorFeed(ctx, xrpcc, author, "", "", 99)
-			if err != nil {
-				return err
-			}
-
-			for i := len(tl.Feed) - 1; i >= 0; i-- {
-				it := tl.Feed[i]
-				if raw {
-					jsonPrint(it)
-				} else {
-					prettyPrintPost(it, uris)
-				}
-			}
-		} else {
-			algo := "reverse-chronological"
-			tl, err := appbsky.FeedGetTimeline(ctx, xrpcc, algo, "", int64(cctx.Int("count")))
-			if err != nil {
-				return err
-			}
-
-			for i := len(tl.Feed) - 1; i >= 0; i-- {
-				it := tl.Feed[i]
-				if raw {
-					jsonPrint(it)
-				} else {
-					prettyPrintPost(it, uris)
-				}
-			}
-		}
-
-		return nil
-
-	},
-}
-
-var actorGetSuggestionsCmd = &cli.Command{
-	Name:      "actorGetSuggestions",
-	ArgsUsage: "[author]",
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-		if err != nil {
-			return err
-		}
-
-		ctx := context.TODO()
-
-		author := cctx.Args().First()
-		if author == "" {
-			author = xrpcc.Auth.Did
-		}
-
-		resp, err := appbsky.ActorGetSuggestions(ctx, xrpcc, "", 100)
-		if err != nil {
-			return err
-		}
-
-		b, err := json.MarshalIndent(resp.Actors, "", "  ")
-		if err != nil {
-			return err
-		}
-
-		fmt.Println(string(b))
-
-		return nil
-
-	},
-}
-
-var feedSetVoteCmd = &cli.Command{
-	Name:      "vote",
-	ArgsUsage: "<post>",
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-		if err != nil {
-			return err
-		}
-
-		arg := cctx.Args().First()
-
-		parts := strings.Split(arg, "/")
-		if len(parts) < 3 {
-			return fmt.Errorf("invalid post uri: %q", arg)
-		}
-		rkey := parts[len(parts)-1]
-		collection := parts[len(parts)-2]
-		did := parts[2]
-
-		fmt.Println(did, collection, rkey)
-		ctx := context.TODO()
-		resp, err := comatproto.RepoGetRecord(ctx, xrpcc, "", collection, did, rkey)
-		if err != nil {
-			return fmt.Errorf("getting record: %w", err)
-		}
-
-		out, err := comatproto.RepoCreateRecord(ctx, xrpcc, &comatproto.RepoCreateRecord_Input{
-			Collection: "app.bsky.feed.like",
-			Repo:       xrpcc.Auth.Did,
-			Record: &lexutil.LexiconTypeDecoder{
-				Val: &appbsky.FeedLike{
-					CreatedAt: time.Now().Format(util.ISO8601),
-					Subject:   &comatproto.RepoStrongRef{Uri: resp.Uri, Cid: *resp.Cid},
-				},
-			},
-		})
-		if err != nil {
-			return fmt.Errorf("creating vote failed: %w", err)
-		}
-		_ = out
-		return nil
-
-	},
 }
 
 var refreshAuthTokenCmd = &cli.Command{
@@ -602,225 +211,6 @@ var refreshAuthTokenCmd = &cli.Command{
 
 		if err := os.WriteFile(cctx.String("auth"), b, 0600); err != nil {
 			return err
-		}
-
-		return nil
-	},
-}
-
-var deletePostCmd = &cli.Command{
-	Name:      "delete",
-	ArgsUsage: `<rkey>`,
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-		if err != nil {
-			return err
-		}
-
-		rkey := cctx.Args().First()
-
-		if rkey == "" {
-			return fmt.Errorf("must specify rkey of post to delete")
-		}
-
-		schema := "app.bsky.feed.post"
-		if strings.Contains(rkey, "/") {
-			parts := strings.Split(rkey, "/")
-			schema = parts[0]
-			rkey = parts[1]
-		}
-
-		return comatproto.RepoDeleteRecord(context.TODO(), xrpcc, &comatproto.RepoDeleteRecord_Input{
-			Repo:       xrpcc.Auth.Did,
-			Collection: schema,
-			Rkey:       rkey,
-		})
-	},
-}
-
-var listAllPostsCmd = &cli.Command{
-	Name: "list",
-	Flags: []cli.Flag{
-		&cli.BoolFlag{
-			Name: "all",
-		},
-		&cli.BoolFlag{
-			Name: "values",
-		},
-		&cli.BoolFlag{
-			Name: "cids",
-		},
-	},
-	ArgsUsage: `<did>|<repo-path>`,
-	Action: func(cctx *cli.Context) error {
-
-		arg := cctx.Args().First()
-		ctx := context.TODO()
-
-		var repob []byte
-		if strings.HasPrefix(arg, "did:") {
-			xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-			if err != nil {
-				return err
-			}
-
-			if arg == "" {
-				arg = xrpcc.Auth.Did
-			}
-
-			rrb, err := comatproto.SyncGetRepo(ctx, xrpcc, arg, "")
-			if err != nil {
-				return err
-			}
-			repob = rrb
-		} else {
-			if len(arg) == 0 {
-				return cli.Exit("must specify DID string or repo path", 127)
-			}
-			fb, err := os.ReadFile(arg)
-			if err != nil {
-				return err
-			}
-
-			repob = fb
-		}
-
-		rr, err := repo.ReadRepoFromCar(ctx, bytes.NewReader(repob))
-		if err != nil {
-			return err
-		}
-
-		collection := "app.bsky.feed.post"
-		if cctx.Bool("all") {
-			collection = ""
-		}
-		vals := cctx.Bool("values")
-		cids := cctx.Bool("cids")
-
-		if err := rr.ForEach(ctx, collection, func(k string, v cid.Cid) error {
-			if !strings.HasPrefix(k, collection) {
-				return repo.ErrDoneIterating
-			}
-
-			fmt.Print(k)
-			if cids {
-				fmt.Println(" - ", v)
-			} else {
-				fmt.Println()
-			}
-			if vals {
-				b, err := rr.Blockstore().Get(ctx, v)
-				if err != nil {
-					return err
-				}
-
-				convb, err := cborToJson(b.RawData())
-				if err != nil {
-					return err
-				}
-				fmt.Println(string(convb))
-			}
-			return nil
-		}); err != nil {
-			return err
-		}
-
-		return nil
-	},
-}
-
-var getNotificationsCmd = &cli.Command{
-	Name:  "notifs",
-	Flags: []cli.Flag{},
-	Action: func(cctx *cli.Context) error {
-		ctx := context.TODO()
-
-		xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-		if err != nil {
-			return err
-		}
-
-		notifs, err := appbsky.NotificationListNotifications(ctx, xrpcc, "", 50, "")
-		if err != nil {
-			return err
-		}
-
-		for _, n := range notifs.Notifications {
-			b, err := json.Marshal(n)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(b))
-		}
-
-		return nil
-	},
-}
-
-var followsCmd = &cli.Command{
-	Name: "follows",
-	Subcommands: []*cli.Command{
-		followsAddCmd,
-		followsListCmd,
-	},
-}
-
-var followsAddCmd = &cli.Command{
-	Name:      "add",
-	Flags:     []cli.Flag{},
-	ArgsUsage: `<user>`,
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, true)
-		if err != nil {
-			return err
-		}
-
-		user := cctx.Args().First()
-
-		follow := appbsky.GraphFollow{
-			LexiconTypeID: "app.bsky.graph.follow",
-			CreatedAt:     time.Now().Format(time.RFC3339),
-			Subject:       user,
-		}
-
-		resp, err := comatproto.RepoCreateRecord(context.TODO(), xrpcc, &comatproto.RepoCreateRecord_Input{
-			Collection: "app.bsky.graph.follow",
-			Repo:       xrpcc.Auth.Did,
-			Record:     &lexutil.LexiconTypeDecoder{&follow},
-		})
-		if err != nil {
-			return err
-		}
-
-		fmt.Println(resp.Uri)
-
-		return nil
-	},
-}
-
-var followsListCmd = &cli.Command{
-	Name:      "list",
-	ArgsUsage: `[actor]`,
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, false)
-		if err != nil {
-			return err
-		}
-
-		user := cctx.Args().First()
-		if user == "" {
-			user = xrpcc.Auth.Did
-		}
-
-		ctx := context.TODO()
-		resp, err := appbsky.GraphGetFollows(ctx, xrpcc, user, "", 100)
-		if err != nil {
-			return err
-		}
-
-		for _, f := range resp.Follows {
-			fmt.Println(f.Did, f.Handle)
 		}
 
 		return nil
@@ -883,66 +273,6 @@ var resetPasswordCmd = &cli.Command{
 			Password: npass,
 			Token:    code,
 		}); err != nil {
-			return err
-		}
-
-		return nil
-	},
-}
-
-var handleCmd = &cli.Command{
-	Name: "handle",
-	Subcommands: []*cli.Command{
-		resolveHandleCmd,
-		updateHandleCmd,
-	},
-}
-
-var resolveHandleCmd = &cli.Command{
-	Name:      "resolve",
-	ArgsUsage: `<handle>`,
-	Action: func(cctx *cli.Context) error {
-		ctx := context.TODO()
-
-		args, err := needArgs(cctx, "handle")
-		if err != nil {
-			return err
-		}
-		handle := args[0]
-
-		phr := &api.ProdHandleResolver{}
-		out, err := phr.ResolveHandleToDid(ctx, handle)
-		if err != nil {
-			return err
-		}
-
-		fmt.Println(out)
-
-		return nil
-	},
-}
-
-var updateHandleCmd = &cli.Command{
-	Name:      "update",
-	ArgsUsage: `<handle>`,
-	Action: func(cctx *cli.Context) error {
-		ctx := context.TODO()
-
-		xrpcc, err := cliutil.GetXrpcClient(cctx, false)
-		if err != nil {
-			return err
-		}
-
-		args, err := needArgs(cctx, "handle")
-		if err != nil {
-			return err
-		}
-		handle := args[0]
-
-		err = comatproto.IdentityUpdateHandle(ctx, xrpcc, &comatproto.IdentityUpdateHandle_Input{
-			Handle: handle,
-		})
-		if err != nil {
 			return err
 		}
 
@@ -1376,40 +706,6 @@ func readDids(f string) ([]string, error) {
 	}
 
 	return out, nil
-}
-
-var syncListReposCmd = &cli.Command{
-	Name: "listRepos",
-	Action: func(cctx *cli.Context) error {
-		xrpcc, err := cliutil.GetXrpcClient(cctx, false)
-		if err != nil {
-			return err
-		}
-
-		var curs string
-		for {
-			out, err := comatproto.SyncListRepos(context.TODO(), xrpcc, curs, 1000)
-			if err != nil {
-				return err
-			}
-
-			if len(out.Repos) == 0 {
-				break
-			}
-
-			for _, r := range out.Repos {
-				fmt.Println(r.Did)
-			}
-
-			if out.Cursor == nil {
-				break
-			}
-
-			curs = *out.Cursor
-		}
-
-		return nil
-	},
 }
 
 func needArgs(cctx *cli.Context, name ...string) ([]string, error) {
